Refactor hello module
