--- hello-go/go.mod
@@ -1,3 +1,0 @@
-module golang-core/hello-go
-
-go 1.23.2
