+++ hello/go.mod
@@ -1,0 +1,3 @@
+module golang-core/hello
+
+go 1.23.2
