+++ hello/greetings/tests/greeting_test.go
@@ -4,7 +4,7 @@ import (
 	"regexp"
 	"testing"
 
-	"golang-core/hello-go/src"
+	"golang-core/hello/greetings"
 )
 
 // TestHelloName calls greetings.Hello with a name, checking
@@ -12,7 +12,7 @@ import (
 func TestHelloName(t *testing.T) {
 	name := "Gladys"
 	want := regexp.MustCompile(`\b` + name + `\b`)
-	msg, err := src.Hello("Gladys")
+	msg, err := greetings.Hello("Gladys")
 	if !want.MatchString(msg) || err != nil {
 		t.Errorf(`Hello("Gladys") = %q, %v, want match for %#q, nil`, msg, err, want)
 	}
@@ -21,7 +21,7 @@ func TestHelloName(t *testing.T) {
 // TestHelloEmpty calls greetings.Hello with an empty string,
 // checking for an error.
 func TestHelloEmpty(t *testing.T) {
-	msg, err := src.Hello("")
+	msg, err := greetings.Hello("")
 	if msg != "" || err == nil {
 		t.Errorf(`Hello("") = %q, %v, want "", error`, msg, err)
 	}
